General file clean-up
